ListPtr.Append now has a version that returns no value
Errors will raise an exception instead of returning False. The old version that returns a Boolean remains, and while it's not deprecated (yet) it is in the crosshairs.

Existing code that uses the CALL keyword to discard the return value WILL need to be updated. Existing code that does not discard the return value will not need to be updated (yet.)