Add ProxyEngine.CA_List
libcurl 7.77.0 added the ability to set the CA list from a memoryblock. This commit adds this capability to the ProxyEngine class. This eliminates the need for a temp file.

This option is supported by the BearSSL (since 7.79.0), mbedTLS (since 7.81.0), rustls (since 7.82.0), OpenSSL (and most variants), Secure Transport and Schannel backends.